+++ include/Game/Camera/CameraRailWatch.hpp
@@ -11,7 +11,7 @@ public:
     virtual CameraTargetObj *calc();
     virtual CamTranslatorBase *createTranslator();
 
-    void setParam(long, long, long, float, float, float);
+    void setParam(s32, s32, s32, f32, f32, f32);
 
     u32 _4C;
     u32 _50;
