+++ include/Game/Camera/CameraShaker.hpp
@@ -27,7 +27,7 @@ public:
     virtual void movement();
 
     void shakeVertical(ESinglyVerticalPower);
-    void shakeInfinity(NameObj *, float, float);
+    void shakeInfinity(NameObj *, f32, f32);
     void stopShakingInfinity(NameObj *);
     void createSinglyVerticalTask();
     void createSinglyHorizontalTask();
@@ -40,7 +40,7 @@ public:
     void addSinglyHorizontalTaskOffset(TVec2f *);
     void addInfinityTaskOffset(TVec2f *);
     void adjustOffsetToScreen(TVec2f *); // TODO
-    void startInfinity(unsigned long, float, float);
+    void startInfinity(u32, f32, f32);
 
     CameraShakeTask *mVerticalTasks[NR_VERTICAL_TASKS];         // 0xC
     CameraShakeTask *mHorizontalTasks[NR_HORIZONTAL_TASKS];     // 0x28
