+++ include/Game/GameAudio/AudTalkSoundData.hpp
@@ -4,7 +4,7 @@
 
 class JAISoundID {
 public:
-    JAISoundID(unsigned long);
+    JAISoundID(u32);
 
     u32 mID;    // 0x0
 };
