+++ include/Game/Util/DirectDraw.hpp
@@ -8,9 +8,9 @@ public:
 
     static void setup(u32, u32, u8);
 
-    static void drawSphere(const TVec3f &, float, unsigned long, unsigned long);
+    static void drawSphere(const TVec3f &, f32, u32, u32);
 
-    static void drawFillFan(const TVec3f &, const TVec3f &, const TVec3f &, unsigned long, float, float, unsigned long);
+    static void drawFillFan(const TVec3f &, const TVec3f &, const TVec3f &, u32, f32, f32, u32);
 
     static void cameraInit3D();
 };
